refactor: move the FeeConsensus into fedimint-core such that it may be reused across all v2 modules
